+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+* Added support to import from linux `perf script` [#135]
+
 ### Fixed
 
 ## [0.6.0]